Check file exists

+++ data/append-remove.sh
@@ -4,21 +4,23 @@ destFile=$1
 textFile=$2
 tempFile=.vimrc-oderint.tmp
 
-cp $destFile $destFile.bak
+if [[ ! -f $destFile ]]; then
+    cp $destFile $destFile.bak
+fi
 
-appendText () {
+appendText() {
     wordCount=$([[ -f $destFile ]] && cat $destFile | wc -w)
 
     ed -s $destFile <<< w
-    
+
     if [[ $wordCount -eq 0 ]]; then
         tail -n +2 $textFile > $destFile
     else
         cat $textFile >> $destFile
-    fi    
+    fi
 }
 
-removeText () {
+removeText() {
     wordCount=$(grep -o "$(cat $textFile)" $destFile | wc -l)
     isTextExists=$([[ $wordCount -ne 0 ]] && echo true || echo false)
